Create accelerometer sensor
