+++ src/view/components/microbit/Microbit.tsx
@@ -4,6 +4,9 @@
 import * as React from "react";
 import "../../styles/Simulator.css";
 import { MicrobitSimulator } from "./MicrobitSimulator";
+import ToolBar from "../toolbar/ToolBar";
+import * as TOOLBAR_SVG from "../../svgs/toolbar_svg";
+import { MICROBIT_TOOLBAR_ID } from "../../components/toolbar/SensorModalUtils";
 
 // Component grouping the functionality for micro:bit functionalities
 
@@ -12,8 +15,23 @@ export class Microbit extends React.Component {
         return (
             <React.Fragment>
                 <MicrobitSimulator />
-                {/* Implement toolbar here */}
+                <ToolBar buttonList={MICROBIT_TOOLBAR_BUTTONS} />
             </React.Fragment>
         );
     }
 }
+
+const MICROBIT_TOOLBAR_BUTTONS: Array<{ label: string; image: JSX.Element }> = [
+    {
+        image: TOOLBAR_SVG.TEMPERATURE_SVG,
+        label: MICROBIT_TOOLBAR_ID.TEMPERATURE,
+    },
+    {
+        image: TOOLBAR_SVG.LIGHT_SVG,
+        label: MICROBIT_TOOLBAR_ID.LIGHT,
+    },
+    {
+        image: TOOLBAR_SVG.MOTION_SVG,
+        label: MICROBIT_TOOLBAR_ID.ACCELEROMETER,
+    },
+];
