+++ src/view/components/toolbar/motion/Accelerometer.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+import { ThreeDimensionSlider } from "./threeDimensionSlider/ThreeDimensionSlider";
+
+export const Accelerometer: React.FC<{}> = () => {
+    return (
+        <div className="AccelerometerBar">
+            <br />
+            <ThreeDimensionSlider />
+        </div>
+    );
+};
